+++ src/main.cpp
@@ -453,6 +453,7 @@ int main(int argc, char* argv[]) {
         compute.init_draw_accum_symmetrize();
         compute.init_smooth();
         compute.init_stroke_smooth();
+        compute.init_mirror_project();
         compute.init_crease();
         compute.init_pinch();
         compute.init_move();